+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 # [1.5.0] - TBA
 - Update FLTK to 1.4.0.
 - Grid and Terminal are no longer experimental.
+- Move Terminal under its own module.
 
 # [1.4.36] - 2024-11-14
 - Add draw::capture_window_part.